Corrected warning.

+++ BOTS/BotsEngine.Damage/Damage.cpp
@@ -57,7 +57,7 @@ const Damage BotsEngine::Damage::operator+(const double value1, const Damage & v
 const Damage BotsEngine::Damage::operator*(const Damage & value1, const double value2)
 {
 	return value2 > 0
-		    ? Damage(round(value1.GetValue() * value2))
+		    ? Damage(static_cast<unsigned int>(round(value1.GetValue() * value2)))
 		    : Damage();
 }
